+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## {{ UNRELEASED_VERSION }} - [{{ UNRELEASED_DATE }}]({{ UNRELEASED_LINK }})
 
 * Added PHP `8.5` support with Terminus `4.2.0`
+* Added warning when `pantheon.yml` sets a `drush_version` below `8` [#354](https://github.com/lando/pantheon/issues/354)
 
 ## v1.13.0 - [March 11, 2026](https://github.com/lando/pantheon/releases/tag/v1.13.0)